+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 ### ⚠ Breaking
 ### ⭐ New Features
 ### 🐞 Bugs Fixed
+
+## v2.0.1 - 2024-05-23
+### ⚠ Breaking
+### ⭐ New Features
+### 🐞 Bugs Fixed
 - removed usage of `org.springframework.util.Base64Utils` since it is deprecated now
 - cookie value was not url-decoded (fixes [#30](https://github.com/sephiroth-j/spring-security-ltpa2-core/issues/30))